+++ pytsite/odm/_queue.py
@@ -1,8 +1,8 @@
-"""
+"""PytSite ODM Queue
 """
 from pymongo import errors as _pymonog_errors
 from bson import errors as _bson_errors
-from pytsite import db as _db, queue as _queue, logger as _logger, cache as _cache
+from pytsite import db as _db, queue as _queue, logger as _logger, cache as _cache, reg as _reg
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -10,6 +10,7 @@ __license__ = 'MIT'
 
 _QUEUE = _queue.Queue('pytsite.odm')
 _ENTITIES_CACHE = _cache.get_pool('pytsite.odm.entities')
+_CACHE_TTL = _reg.get('odm.cache.ttl')
 
 
 def _entity_save(args: dict):
@@ -29,7 +30,7 @@ def _entity_save(args: dict):
             collection.replace_one({'_id': fields_data['_id']}, fields_data)
 
         # Update cache
-        _ENTITIES_CACHE.put('{}.{}'.format(args['model'], fields_data['_id']), fields_data)
+        _ENTITIES_CACHE.put('{}.{}'.format(args['model'], fields_data['_id']), fields_data, _CACHE_TTL)
 
     except (_bson_errors.BSONError, _pymonog_errors.PyMongoError) as e:
         _logger.error(e)